Always install both init script and systemd service
Currently configure.ac checks if systemd is installed, and
generates a Makefile that does not install systemd service
if it's built on a system without systemd. Ideally this
should be done on package's post-install phase, but it's
not necessary since init scripts and systemd service can
live together.

This patch adds a default/hardcoded value for systemd
directory, which makes its service is installed on default
place even if configured on a system withtout systemd.

It also changes Makefiles to always install both init
script and systemd service.

Signed-off-by: Lucio Correia <luciojhc@linux.vnet.ibm.com>